requirements.txt
Flask==3.1.0
geopy==2.4.1
langchain==0.3.25
langchain_community==0.3.23
langchain_core==0.3.59
langchain_google_genai==2.1.4
neo4j==5.28.1
numpy==2.2.5
pandas==2.2.3
protobuf==6.30.2
pydantic==2.11.4
python-dotenv==1.1.0
scikit_learn==1.6.1
sentence_transformers==2.2.2
